backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.60.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
bitarray==3.6.0
certifi==2025.7.14
charset-normalizer==3.4.2
ckzg==1.0.2
click==8.2.1
cytoolz==1.0.1
dataclasses-json==0.6.7
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
eth-account==0.11.3
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==1.0.1
eth-typing==4.4.0
eth-utils==4.1.1
eth_abi==5.2.0
executing==2.2.0
fastapi==0.116.1
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.7.0
gotrue==2.12.3
greenlet==3.2.3
grpcio==1.74.0
h11==0.16.0
h2==4.2.0
hexbytes==0.3.1
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
hyperframe==6.1.0
idna==3.10
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
langchain==0.3.27
langchain-anthropic==0.3.18
langchain-community==0.3.27
langchain-core==0.3.72
langchain-huggingface==0.3.1
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langgraph==0.6.2
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.2
langgraph-sdk==0.2.0
langsmith==0.4.10
lru-dict==1.2.0
marshmallow==3.26.1
matplotlib-inline==0.1.7
multidict==6.6.3
mypy_extensions==1.1.0
numpy==2.3.2
openai==1.98.0
orjson==3.11.1
ormsgpack==1.10.0
packaging==25.0
parsimonious==0.10.0
parso==0.8.4
pexpect==4.9.0
portalocker==3.2.0
postgrest==1.1.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-solc-x==2.0.4
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.13.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyunormalize==16.0.0
PyYAML==6.0.2
qdrant-client==1.15.1
realtime==2.6.0
redis==6.2.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rlp==4.1.0
rpds-py==0.27.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
stack-data==0.6.3
starlette==0.47.2
storage3==0.12.0
StrEnum==0.4.15
supabase==2.17.0
supafunc==0.10.1
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.4
toolz==1.0.0
tqdm==4.67.1
traitlets==5.14.3
tree-sitter==0.25.0
tree-sitter-solidity==1.2.12
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
web3==6.20.4
websockets==13.1
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
